src/main/java/school/faang/user_service/service/user/UserService.java
package school.faang.user_service.service.user;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;
import school.faang.user_service.dto.event.EventDto;
import school.faang.user_service.entity.User;
import school.faang.user_service.exception.DataValidationException;
import school.faang.user_service.repository.UserRepository;

@Component
@RequiredArgsConstructor
public class UserService {
    private final UserRepository userRepository;

    public User findOwnerById(EventDto event) {
        return userRepository.findById(event.getOwnerId()).orElseThrow(() -> new DataValidationException("User not found"));
    }
}
